[HK] reorganizing catalog tests
