+++ plugins/org.locationtech.udig.catalog.tests/src/org/locationtech/udig/catalog/tests/internal/LessInterestingService.java
@@ -1,3 +1,13 @@
+/**
+ * uDig - User Friendly Desktop Internet GIS client
+ * https://locationtech.org/projects/technology.udig
+ * (C) 2015, Eclipse Foundation
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * (http://www.eclipse.org/legal/epl-v10.html), and the Eclipse Distribution
+ * License v1.0 (http://www.eclipse.org/org/documents/edl-v10.html).
+ */
 package org.locationtech.udig.catalog.tests.internal;
 
 import java.io.IOException;
@@ -12,11 +22,11 @@ import org.locationtech.udig.catalog.IGeoResource;
 import org.locationtech.udig.catalog.IService;
 import org.locationtech.udig.catalog.IServiceInfo;
 import org.locationtech.udig.catalog.ServiceExtension;
+import org.locationtech.udig.catalog.URLUtilsTest;
 
 public class LessInterestingService implements ServiceExtension {
 
-    public class LessInterestingServiceImpl extends IService
-    {
+    public class LessInterestingServiceImpl extends IService {
         private URL id;
 
         public LessInterestingServiceImpl(URL id) {
@@ -34,8 +44,7 @@ public class LessInterestingService implements ServiceExtension {
         }
 
         @Override
-        public List<? extends IGeoResource> resources(IProgressMonitor monitor)
-                throws IOException {
+        public List<? extends IGeoResource> resources(IProgressMonitor monitor) throws IOException {
             return null;
         }
 
@@ -47,41 +56,40 @@ public class LessInterestingService implements ServiceExtension {
                 public double getMetric() {
                     return 0.0;
                 }
-                
+
             };
         }
-        
+
         @Override
         public IServiceInfo getInfo(IProgressMonitor monitor) throws IOException {
             return createInfo(monitor);
         }
-        
+
         @Override
         public Map<String, Serializable> getConnectionParams() {
             return null;
         }
     }
-    
+
     @Override
     public IService createService(URL id, Map<String, Serializable> params) {
         URL testURL = null;
         try {
-            testURL = new URL(CatalogImplTest.SERVICE_COMPARISON_TEST_URL);
+            testURL = new URL(URLUtilsTest.SERVICE_COMPARISON_TEST_URL);
+        } catch (Exception e) {
         }
-        catch(Exception e) {}
         Serializable paramId = params.get("id"); //$NON-NLS-1$
-        if( paramId == null || !paramId.equals(testURL))
-        {
+        if (paramId == null || !paramId.equals(testURL)) {
             return null;
         }
-        
+
         return new LessInterestingServiceImpl((URL) params.get("id")); //$NON-NLS-1$
     }
 
     @Override
     public Map<String, Serializable> createParams(URL url) {
-        Map<String, Serializable> params = new TreeMap<String, Serializable>();
-        params.put("id",url); //$NON-NLS-1$
+        Map<String, Serializable> params = new TreeMap<>();
+        params.put("id", url); //$NON-NLS-1$
         return params;
     }
